add emitProgress to WSEventEmitter
