+++ workers/loc.api/ws-transport/ws.event.emitter.js
@@ -1,0 +1,23 @@
+'use strict'
+
+let deps = {}
+
+class WSEventEmitter {
+  constructor (inDeps) {
+    this.deps = { ...deps, ...inDeps }
+  }
+
+  static inject (inDeps) {
+    deps = { ...deps, ...inDeps }
+  }
+
+  emitProgress (handler = () => {}) {
+    return this.deps.wsTransport.send(
+      handler,
+      'emitProgress',
+      { deps: this.deps }
+    )
+  }
+}
+
+module.exports = WSEventEmitter
