--- workers/loc.api/ws-transport/ws.event.emmiter.js
@@ -1,19 +1,0 @@
-'use strict'
-
-let deps = {}
-
-class WSEventEmmiter {
-  constructor (inDeps) {
-    this.deps = { ...deps, ...inDeps }
-  }
-
-  static inject (inDeps) {
-    deps = { ...deps, ...inDeps }
-  }
-
-  emmitProgress (handler = () => {}) {
-    return this.deps.wsTransport.send(handler, 'progress')
-  }
-}
-
-module.exports = WSEventEmmiter
